% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vpa.R
\name{b}
\alias{b}
\title{Return biomass of given year}
\usage{
b(vpares, year, unit = "kton", add_jpdigit = TRUE, draft = FALSE)
}
\arguments{
\item{year}{Year of the value you want}

\item{add_jpdigit}{If TRUE, Japanese kanji digit mark will be added}

\item{draft}{If TRUE, value will be returned wrapped by HTML tag
for highlight in document review}
}
\description{
Return biomass of given year
}
