% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_tool.R
\name{checkme}
\alias{checkme}
\title{Highlight hand-written phrase for review}
\usage{
checkme(string, draft)
}
\arguments{
\item{string}{String to be highlighted}

\item{draft}{If FALSE, \code{string} will be returned without highlight}
}
\description{
Highlight hand-written phrase for review
}
