% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_decimals.R
\name{conv.catch_kg}
\alias{conv.catch_kg}
\title{Convert catch in kilogram to various decimal order}
\usage{
\method{conv}{catch_kg}(catch_kg, to, add_jpdigit = TRUE)
}
\arguments{
\item{catch_kg}{Catch in kilogram}

\item{to}{decimal order to convert in character such as
\describe{
  \item{"ton"}{ton}
  \item{"dton"}{10 ton}
  \item{"cton"}{100 ton}
  \item{"kton"}{kilo ton}
  \item{"10kton"}{10 kilo ton}
}}

\item{add_jpdigit}{If TRUE, Japanese kanji digit mark will be added}
}
\description{
Convert catch in kilogram to various decimal order
}
