% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_decimals.R
\name{conv}
\alias{conv}
\title{Convert decimals}
\usage{
conv(value, ...)
}
\arguments{
\item{value}{Value to be converted}

\item{...}{Arguments to be passed to methods}
}
\description{
Convert decimals
}
