% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_decimals.R
\name{conv.recruit_106fish}
\alias{conv.recruit_106fish}
\title{Convert number of recruitments in 10^6 fishes to several decimal order}
\usage{
\method{conv}{recruit_106fish}(r_106fish, to, add_jpdigit = TRUE)
}
\arguments{
\item{r_106fish}{Number of recruitments in "million fish"}

\item{to}{decimal order to convert in character such as
\describe{
  \item{"mil"}{million fishes}
  \item{"100mil"}{100 million fishes, "oku" in Japanese}
  #' }}

\item{add_jpdigit}{If TRUE, Japanese kanji digit mark will be added}
}
\description{
Convert number of recruitments in 10^6 fishes to several decimal order
}
