% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vpa.R
\name{faa}
\alias{faa}
\title{Return fishing mortality of given year}
\usage{
faa(vpares, year, draft = FALSE)
}
\arguments{
\item{year}{Year of the value you want}

\item{draft}{If TRUE, value will be returned wrapped by HTML tag
for highlight in document review}
}
\description{
Return fishing mortality of given year
}
