% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vpa.R
\name{harvest_rate}
\alias{harvest_rate}
\title{Return harvest rate of given year}
\usage{
harvest_rate(vpares, year, draft = FALSE)
}
\arguments{
\item{draft}{If FALSE, \code{string} will be returned without highlight}
}
\description{
Return harvest rate of given year
}
