% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_tool.R
\name{highlight}
\alias{highlight}
\title{Highlight string for review}
\usage{
highlight(string, draft, color = "skyblue")
}
\arguments{
\item{string}{String to be highlighted}

\item{draft}{If FALSE, \code{string} will be returned without highlight}

\item{color}{Color for highlighting}
}
\description{
Suitable for strings generated by eval because they will be kept "inline"
}
