% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vpa.R
\name{pull_value_from_vpa_result}
\alias{pull_value_from_vpa_result}
\title{Pull weight value from vpa result}
\usage{
pull_value_from_vpa_result(vpares, yr, varname)
}
\arguments{
\item{yr}{Year to filter output}

\item{varname}{Name of variable from \code($stat) of
frasyr::convert_vpa_tibble output}
}
\description{
Pull weight value from vpa result
}
