% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_class}
\alias{set_class}
\title{Set given class to object}
\usage{
set_class(x, class)
}
\arguments{
\item{x}{Object to be given \code{class}}

\item{class}{Class to be given \code{x}}
}
\description{
To use S3 function
}
