% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_range}
\alias{str_range}
\title{Concatenate 2-length-vector with specific character}
\usage{
str_range(range_vec, sep = "–")
}
\arguments{
\item{range_vec}{Vector composed of two objects}

\item{sep}{Character between two vectors}
}
\description{
Concatenate 2-length-vector with specific character
}
