% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_decimals.R
\name{toggle_paste}
\alias{toggle_paste}
\title{Create character from numbers
Concatenate two characters}
\usage{
toggle_paste(x, y, ..., sep = "", concat = TRUE)
}
\arguments{
\item{x}{The first character}

\item{y}{The other}

\item{...}{Additional arguments to be concatenated}

\item{sep}{Separate character between two args}

\item{concat}{If FALSE, x will be returned}
}
\description{
Create character from numbers
Concatenate two characters
}
